refactor: extract mcp stdio test support
