+++ crates/rulemorph_mcp/tests/stdio.rs
@@ -1,103 +1,12 @@
 use std::fs;
-use std::io::{BufRead, BufReader, Write};
-use std::path::{Path, PathBuf};
-use std::process::{Child, ChildStdin, Command, Stdio};
 
 use rulemorph::parse_rule_file;
 use serde_json::{Value, json};
 use tempfile::tempdir;
 
-struct McpServer {
-    child: Child,
-    stdin: Option<ChildStdin>,
-    stdout: BufReader<std::process::ChildStdout>,
-}
+mod common;
 
-impl McpServer {
-    fn start() -> Self {
-        let bin = env!("CARGO_BIN_EXE_rulemorph-mcp");
-        let mut child = Command::new(bin)
-            .env("RULEMORPH_MCP_ALLOW_ANY_PATH", "1")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::inherit())
-            .spawn()
-            .expect("spawn mcp server");
-
-        let stdin = child.stdin.take().expect("take stdin");
-        let stdout = child.stdout.take().expect("take stdout");
-
-        Self {
-            child,
-            stdin: Some(stdin),
-            stdout: BufReader::new(stdout),
-        }
-    }
-
-    fn start_with_allowed_root(root: &std::path::Path) -> Self {
-        let bin = env!("CARGO_BIN_EXE_rulemorph-mcp");
-        let mut child = Command::new(bin)
-            .env("RULEMORPH_MCP_ALLOWED_ROOTS", root)
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::inherit())
-            .spawn()
-            .expect("spawn mcp server");
-
-        let stdin = child.stdin.take().expect("take stdin");
-        let stdout = child.stdout.take().expect("take stdout");
-
-        Self {
-            child,
-            stdin: Some(stdin),
-            stdout: BufReader::new(stdout),
-        }
-    }
-
-    fn send(&mut self, message: &Value) -> Value {
-        let text = serde_json::to_string(message).expect("serialize request");
-        let stdin = self.stdin.as_mut().expect("stdin available");
-        writeln!(stdin, "{}", text).expect("write request");
-        stdin.flush().expect("flush request");
-
-        let mut line = String::new();
-        self.stdout.read_line(&mut line).expect("read response");
-        assert!(!line.trim().is_empty(), "empty response");
-        serde_json::from_str(&line).expect("parse response")
-    }
-
-    fn shutdown(mut self) {
-        self.stdin.take();
-        let _ = self.child.wait();
-    }
-}
-
-fn initialize(server: &mut McpServer) {
-    let request = json!({
-        "jsonrpc": "2.0",
-        "id": 1,
-        "method": "initialize",
-        "params": {
-            "protocolVersion": "2024-11-05",
-            "capabilities": {},
-            "clientInfo": {
-                "name": "tests",
-                "version": "0.0"
-            }
-        }
-    });
-    let response = server.send(&request);
-    assert_eq!(response["result"]["protocolVersion"], "2024-11-05");
-}
-
-fn core_fixtures_dir() -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .expect("workspace crates dir")
-        .join("rulemorph")
-        .join("tests")
-        .join("fixtures")
-}
+use common::stdio::{McpServer, core_fixtures_dir, initialize};
 
 #[test]
 fn initialize_and_list_tools() {
